api: refactor species formatting into a service
